feat(wasm): unblock streams in the browser
