+++ xmtp_api_http/Cargo.toml
@@ -18,6 +18,7 @@ thiserror = "2.0"
 tokio = { workspace = true, features = ["sync", "rt", "macros"] }
 xmtp_proto = { path = "../xmtp_proto", features = ["proto_full"] }
 async-trait = "0.1"
+bytes = "1.9"
 
 [dev-dependencies]
 xmtp_proto = { path = "../xmtp_proto", features = ["test-utils"] }
