+++ xmtp_api_http/src/util.rs
@@ -1,9 +1,5 @@
-use futures::{
-    stream::{self, StreamExt},
-    Stream,
-};
+use crate::http_stream::SubscriptionItem;
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
-use serde_json::Deserializer;
 use std::io::Read;
 use xmtp_proto::{Error, ErrorKind};
 
@@ -23,11 +19,6 @@ pub(crate) struct ErrorResponse {
     details: Vec<String>,
 }
 
-#[derive(Deserialize, Serialize, Debug)]
-pub(crate) struct SubscriptionItem<T> {
-    pub result: T,
-}
-
 /// handle JSON response from gRPC, returning either
 /// the expected deserialized response object or a gRPC [`Error`]
 pub fn handle_error<R: Read, T>(reader: R) -> Result<T, Error>
@@ -40,78 +31,6 @@ where
         Ok(GrpcResponse::Empty {}) => Ok(Default::default()),
         Ok(GrpcResponse::SubscriptionItem(item)) => Ok(item.result),
         Err(e) => Err(Error::new(ErrorKind::QueryError).with(e.to_string())),
-    }
-}
-
-#[cfg(target_arch = "wasm32")]
-pub fn create_grpc_stream<
-    T: Serialize + Send + 'static,
-    R: DeserializeOwned + Send + std::fmt::Debug + 'static,
->(
-    request: T,
-    endpoint: String,
-    http_client: reqwest::Client,
-) -> stream::LocalBoxStream<'static, Result<R, Error>> {
-    create_grpc_stream_inner(request, endpoint, http_client).boxed_local()
-}
-
-#[cfg(not(target_arch = "wasm32"))]
-pub fn create_grpc_stream<
-    T: Serialize + Send + 'static,
-    R: DeserializeOwned + Send + std::fmt::Debug + 'static,
->(
-    request: T,
-    endpoint: String,
-    http_client: reqwest::Client,
-) -> stream::BoxStream<'static, Result<R, Error>> {
-    create_grpc_stream_inner(request, endpoint, http_client).boxed()
-}
-
-pub fn create_grpc_stream_inner<
-    T: Serialize + Send + 'static,
-    R: DeserializeOwned + Send + std::fmt::Debug + 'static,
->(
-    request: T,
-    endpoint: String,
-    http_client: reqwest::Client,
-) -> impl Stream<Item = Result<R, Error>> {
-    async_stream::stream! {
-        let request = http_client
-                .post(endpoint)
-                .json(&request)
-                .send()
-                .await
-                .map_err(|e| Error::new(ErrorKind::MlsError).with(e))?;
-
-        let mut remaining = vec![];
-        for await bytes in request.bytes_stream() {
-            let bytes = bytes
-                .map_err(|e| Error::new(ErrorKind::SubscriptionUpdateError).with(e.to_string()))?;
-            let bytes = &[remaining.as_ref(), bytes.as_ref()].concat();
-            let de = Deserializer::from_slice(bytes);
-            let mut stream = de.into_iter::<GrpcResponse<R>>();
-            'messages: loop {
-                let response = stream.next();
-                let res = match response {
-                    Some(Ok(GrpcResponse::Ok(response))) => Ok(response),
-                    Some(Ok(GrpcResponse::SubscriptionItem(item))) => Ok(item.result),
-                    Some(Ok(GrpcResponse::Err(e))) => {
-                        Err(Error::new(ErrorKind::MlsError).with(e.message))
-                    }
-                    Some(Err(e)) => {
-                        if e.is_eof() {
-                            remaining = (&**bytes)[stream.byte_offset()..].to_vec();
-                            break 'messages;
-                        } else {
-                            Err(Error::new(ErrorKind::MlsError).with(e.to_string()))
-                        }
-                    }
-                    Some(Ok(GrpcResponse::Empty {})) => continue 'messages,
-                    None => break 'messages,
-                };
-                yield res;
-            }
-        }
     }
 }
